requirements.txt
pandas==2.1.3
PyYAML==6.0
python-dotenv==1.0.0
flask
Requests==2.31.0

functions_framework==3.5.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.3
google-cloud-pubsub==2.18.4

openmeteo_requests==1.1.0
requests_cache==1.1.1
retry_requests==2.0.0

#discord.py==2.3.2



